pyproject.toml
[project]
name = "solid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.14.1",
    "curl-cffi>=0.14.0",
    "fastapi>=0.128.0",
    "google-genai>=1.62.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.10.1",
    "ruff>=0.14.14",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.36.0",
]
